organized review and user route files
