+++ server/api.js
@@ -4,11 +4,11 @@ const db = require('APP/db')
 const api = module.exports = require('express').Router()
 
 api
-  .use('/auth', require('./auth'))
-  .use('/users', require('./users'))
+  .use('/auth', require('./routes/users/auth'))
+  .use('/users', require('./routes/users/users'))
   .use('/products', require('./routes/products/product-router'))
   .use('/cart', require('./routes/orders/order-router'))
-  .use('/reviews', require('./routes/review-router') )
+  .use('/reviews', require('./routes/reviews/review-router') )
 
 // Send along any errors
 api.use((err, req, res, next) => {
